#kentprojects-64 Filled out addStaff() method and modified User_Year_Map to use INT instead of BIGINT

+++ database/tables/002-User/20-User_Year_Map.sql
@@ -4,7 +4,7 @@
  * @link: http://kentprojects.com
  */
 CREATE TABLE IF NOT EXISTS `User_Year_Map` (
-	`user_id` BIGINT UNSIGNED AUTO_INCREMENT NOT NULL,
+	`user_id` INT UNSIGNED AUTO_INCREMENT NOT NULL,
 	`year` INT(4) UNSIGNED NOT NULL,
 	PRIMARY KEY (`user_id`, `year`),
 	FOREIGN KEY (`user_id`) REFERENCES `User` (`user_id`) ON UPDATE CASCADE ON DELETE CASCADE,
